delete-question use case done!

+++ src/domain/forum/application/use-cases/delete-question.ts
@@ -2,6 +2,7 @@ import { QuestionsRepository } from "../repositories/questions-repository";
 
 interface IDeleteQuestionUseCaseProps {
   questionId: string;
+  authorId: string;
 }
 
 interface IDeleteQuestionUseCaseResponse {}
@@ -11,11 +12,16 @@ export class DeleteQuestionUseCase {
 
   async execute({
     questionId,
+    authorId,
   }: IDeleteQuestionUseCaseProps): Promise<IDeleteQuestionUseCaseResponse> {
     const question = await this.questionsRepository.findById(questionId);
 
     if (!question) {
       throw new Error("Question not found.");
+    }
+
+    if (authorId !== question.authorId.toString()) {
+      throw new Error("Not Allowed.");
     }
 
     await this.questionsRepository.delete(question);
